[VULN-1021929] Replacing NSCoding with NSSecureCoding
Using NSSecureCoding instead of NSCoding ensures that we are decoding objects to the correct type. This should prevent a class of attacks based on passing malicious JSON that when decoded will decode into some other object with additional code or trigger memory corruption allowing for exploitation. As a part of this we also changed `NSKeyedArchiver` in `PINDiskCache` to set `YES` for `requiringSecureCoding`  that way it will make sure that any objects being encoded conform to `NSSecureCoding` as well.

For more details on NSSecureCoding see: https://nshipster.com/nssecurecoding/